eventgo-backend/party-booking-service/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-dotenv==1.0.0
requests==2.32.3